Remove / break up within-loop words
The inner loop words (I...) hid a lot of underlying complexity in how solutions were calculated. I've removed them for saddle-points, and broken them into more simpler words for mixed strategies. As part of trying to make row and column labels be handled in the same way, I've moved V calculation to after the P1 and P2 strategies are calculated, instead of doing them at the same time. This will slow things down a bit, but makes the code easier to factor.